Add initial static runtime support
