+++ compiler+runtime/bin/ar-merge
@@ -8,6 +8,40 @@ ar="@CMAKE_AR@"
 cmake_binary_dir="@CMAKE_CURRENT_BINARY_DIR@"
 cache_dir="@CMAKE_CURRENT_BINARY_DIR@/ar-cache"
 
+# $1 = subdir
+function find_objects
+{
+  find "${cache_dir}/common" -type f -exec cat {} +
+  find "${cache_dir}/${1}" -type f -exec cat {} +
+}
+
+# $1 = static archive name
+function subdir
+{
+  case "${1}" in
+    libjank-dynamic.a|libjank-compiler.a|libjankzip.a)
+      echo "dynamic"
+      ;;
+    libjank-static.a)
+      echo "static"
+      ;;
+    *)
+      echo "common"
+      ;;
+  esac
+}
+
+# $1 = object files
+# $2 = output static lib
+function build_archive
+{
+  object_files="${1}"
+  merge_output="${2}"
+  rm -f "${merge_output}"
+  # shellcheck disable=SC2086
+  "${ar}" qc "${merge_output}" ${object_files}
+}
+
 command="${1}"
 shift
 
@@ -21,22 +55,26 @@ case "${command}" in
     ;;
 
   # We use this prior to building the jank executable, but after building all libraries.
-  merge)
-    object_files="$(find ${cache_dir} -type f -exec cat {} +)"
-    merge_output="${cmake_binary_dir}/libjank-standalone-phase-1.a"
-    rm -f "${merge_output}"
-    # shellcheck disable=SC2086
-    "${ar}" qc "${merge_output}" ${object_files}
+  merge-phase-1)
+    object_files="$(find_objects dynamic)"
+    merge_output="${cmake_binary_dir}/libjank-dynamic-runtime-phase-1.a"
+    build_archive "${object_files}" "${merge_output}"
+
+    object_files="$(find_objects static)"
+    merge_output="${cmake_binary_dir}/libjank-static-runtime-phase-1.a"
+    build_archive "${object_files}" "${merge_output}"
     ;;
 
   # This follows merge, but also bundles in the Clojure core library object files, following
   # jank's phase 2 building.
   merge-phase-2)
-    object_files="$(find ${cache_dir} -type f -exec cat {} +) @jank_clojure_core_binary_output@ @jank_nrepl_server_outputs_str@"
-    merge_output="${cmake_binary_dir}/libjank-standalone.a"
-    rm -f "${merge_output}"
-    # shellcheck disable=SC2086
-    "${ar}" qc "${merge_output}" ${object_files}
+    object_files="$(find_objects dynamic) @jank_clojure_core_binary_output@ @jank_nrepl_server_outputs_str@"
+    merge_output="${cmake_binary_dir}/libjank-dynamic-runtime.a"
+    build_archive "${object_files}" "${merge_output}"
+
+    object_files="$(find_objects static) @jank_clojure_core_binary_output@ @jank_nrepl_server_outputs_str@"
+    merge_output="${cmake_binary_dir}/libjank-static-runtime.a"
+    build_archive "${object_files}" "${merge_output}"
     ;;
 
   # No custom command, so just pass this onto the original AR but keep track of all of
@@ -45,9 +83,11 @@ case "${command}" in
     output="${1}"
     shift
     base="${output//\//_}"
-    rm -f "${cache_dir}/${base}.list"
-    mkdir -p "$(dirname ${cache_dir}/${base}.list)"
-    echo "$@" > "${cache_dir}/${base}.list"
+    dir="${cache_dir}/$(subdir "${output}")"
+    mkdir -p "${dir}"
+    rm -f "${dir}/${base}.list"
+    mkdir -p "$(dirname "${dir}/${base}.list")"
+    echo "$@" > "${dir}/${base}.list"
     "${ar}" "${command}" "${output}" "$@"
     ;;
 esac
