+++ compiler+runtime/include/cpp/jank/codegen/api.hpp
@@ -23,6 +23,9 @@ jank::runtime::obj::symbol_ref
 _jank_symbol(jank::runtime::object_ref const meta, char const * const ns, char const * const name);
 jank::runtime::obj::persistent_string_ref _jank_string(char const * const s);
 jank::runtime::obj::jit_function_ref _jank_fn(jank::runtime::callable_arity_flags const flags);
+jank::runtime::obj::persistent_vector_ref _jank_vec(jank::u64 const elems, ...);
+jank::runtime::obj::persistent_array_map_ref _jank_amap(jank::u64 const pairs, ...);
+jank::runtime::obj::persistent_hash_map_ref _jank_hmap(jank::u64 const pairs, ...);
 jank::runtime::obj::jit_variadic_function_ref
 _jank_vfn(jank::runtime::callable_arity_flags const flags);
 jank::runtime::obj::jit_closure_ref
