+++ compiler+runtime/include/cpp/jank/runtime/context.hpp
@@ -58,7 +58,6 @@ namespace jank::runtime
     /* Adds the current ns to unqualified symbols and resolves the ns of qualified symbols.
      * Does not intern. */
     obj::symbol_ref qualify_symbol(obj::symbol_ref const) const;
-    jtl::option<object_ref> find_local(obj::symbol_ref const);
 
     jtl::result<var_ref, jtl::immutable_string> intern_var(obj::symbol_ref const qualified_name);
     jtl::result<var_ref, jtl::immutable_string> intern_var(jtl::immutable_string const &);
