+++ compiler+runtime/src/cpp/jank/codegen/api_static.cpp
@@ -1,0 +1,6 @@
+#include <jank/codegen/api.hpp>
+
+jank::runtime::object_ref _jank_eval_str(char const *)
+{
+  return jank::runtime::jank_nil;
+}
