+++ compiler+runtime/src/cpp/jank/codegen/cpp_processor.cpp
@@ -245,26 +245,20 @@ namespace jank::codegen
               util::format_to(buffer, "jank::runtime::jank_false");
             }
           }
-          else if constexpr(std::same_as<T, obj::integer>)
+          else if constexpr(jtl::is_any_same<T, obj::integer, obj::small_integer>)
           {
-            util::format_to(buffer, "_jank_int({})", typed_o->data);
-          }
-          else if constexpr(std::same_as<T, obj::small_integer>)
-          {
-            util::format_to(buffer, "_jank_small_int({})", typed_o->data);
-          }
-          else if constexpr(jtl::is_any_same<T, obj::real, obj::small_real>)
-          {
-            if constexpr(std::same_as<T, obj::real>)
+            if(static_cast<i32>(typed_o->data) == typed_o->data)
             {
-              util::format_to(buffer, "_jank_real(");
+              util::format_to(buffer, "_jank_small_int({})", typed_o->data);
             }
             else
             {
-              util::format_to(buffer, "_jank_small_real(");
+              util::format_to(buffer, "_jank_int({})", typed_o->data);
             }
-
-            util::format_to(buffer, "{})", typed_o->data);
+          }
+          else if constexpr(jtl::is_any_same<T, obj::real, obj::small_real>)
+          {
+            util::format_to(buffer, "_jank_small_real({})", typed_o->data);
           }
           else if constexpr(std::same_as<T, obj::big_integer>)
           {
@@ -349,20 +343,25 @@ namespace jank::codegen
             }
             else
             {
-              util::format_to(buffer,
-                              "jank::runtime::make_box<jank::runtime::obj::persistent_vector>(");
-              if(should_gen_meta(typed_o->meta))
+              auto const has_meta{ should_gen_meta(typed_o->meta) };
+              if(has_meta)
               {
-                gen_constant(typed_o->meta, buffer);
-                util::format_to(buffer, ",");
+                util::format_to(buffer, "jank::runtime::reset_meta(");
               }
-              util::format_to(buffer, "std::in_place ");
-              for(auto const &form : typed_o->data)
+              util::format_to(buffer, "_jank_vec({} ", typed_o->count());
+              for(auto const e : typed_o->data)
               {
                 util::format_to(buffer, ", ");
-                gen_constant(form, buffer);
+                gen_constant(e, buffer);
+                util::format_to(buffer, ".erase()");
               }
               util::format_to(buffer, ")");
+              if(has_meta)
+              {
+                util::format_to(buffer, ",");
+                gen_constant(typed_o->meta, buffer);
+                util::format_to(buffer, ")");
+              }
             }
           }
           else if constexpr(std::same_as<T, obj::persistent_list>)
@@ -439,31 +438,27 @@ namespace jank::codegen
             }
             else
             {
-              bool need_comma{};
-              if(should_gen_meta(typed_o->meta))
+              auto const has_meta{ should_gen_meta(typed_o->meta) };
+              if(has_meta)
               {
-                util::format_to(
-                  buffer,
-                  "jank::runtime::obj::persistent_array_map::create_unique_with_meta(");
-                gen_constant(typed_o->meta, buffer);
-                need_comma = true;
+                util::format_to(buffer, "jank::runtime::reset_meta(");
               }
-              else
+              util::format_to(buffer, "_jank_amap({} ", typed_o->count());
+              for(auto const p : typed_o->data)
               {
-                util::format_to(buffer, "jank::runtime::obj::persistent_array_map::create_unique(");
-              }
-              for(auto const &form : typed_o->data)
-              {
-                if(need_comma)
-                {
-                  util::format_to(buffer, ", ");
-                }
-                need_comma = true;
-                gen_constant(form.first, buffer);
                 util::format_to(buffer, ", ");
-                gen_constant(form.second, buffer);
+                gen_constant(p.first, buffer);
+                util::format_to(buffer, ".erase(), ");
+                gen_constant(p.second, buffer);
+                util::format_to(buffer, ".erase()");
               }
               util::format_to(buffer, ")");
+              if(has_meta)
+              {
+                util::format_to(buffer, ",");
+                gen_constant(typed_o->meta, buffer);
+                util::format_to(buffer, ")");
+              }
             }
           }
           else if constexpr(std::same_as<T, obj::persistent_hash_map>)
@@ -483,15 +478,23 @@ namespace jank::codegen
               auto const has_meta{ should_gen_meta(typed_o->meta) };
               if(has_meta)
               {
-                util::format_to(buffer, "jank::runtime::with_meta(");
+                util::format_to(buffer, "jank::runtime::reset_meta(");
               }
-              util::format_to(buffer,
-                              "_jank_eval_str(\"{}\")",
-                              util::escape(typed_o->to_code_string()));
+              util::format_to(buffer, "_jank_hmap({} ", typed_o->count());
+              for(auto const p : typed_o->data)
+              {
+                util::format_to(buffer, ", ");
+                gen_constant(p.first, buffer);
+                util::format_to(buffer, ".erase(), ");
+                gen_constant(p.second, buffer);
+                util::format_to(buffer, ".erase()");
+              }
+              util::format_to(buffer, ")");
               if(has_meta)
               {
                 util::format_to(buffer, ",");
                 gen_constant(typed_o->meta, buffer);
+                util::format_to(buffer, ")");
               }
             }
           }
