+++ compiler+runtime/src/cpp/jank/jit/processor_static.cpp
@@ -1,0 +1,12 @@
+#include <jank/jit/processor.hpp>
+
+namespace jank::jit
+{
+  processor::processor(jtl::immutable_string const &)
+  {
+  }
+
+  processor::~processor()
+  {
+  }
+}
