+++ compiler+runtime/src/cpp/jank/runtime/ns_static.cpp
@@ -1,0 +1,9 @@
+#include <jank/runtime/ns.hpp>
+
+namespace jank::runtime
+{
+  jtl::result<void, error_ref> ns::refer_global(object_ref const)
+  {
+    return ok();
+  }
+}
