+++ compiler+runtime/src/cpp/jank/util/environment_static.cpp
@@ -1,0 +1,10 @@
+#include <jank/util/environment.hpp>
+
+namespace jank::util
+{
+  jtl::immutable_string const &binary_version()
+  {
+    static jtl::immutable_string res{ "static" };
+    return res;
+  }
+}
